franky makes a massive refactor to create a GFX api abstraction. this is a twist on the PIMPL. i did not want to lose the inline and add a bunch of virtual calls in something that potentially require all the perormance that it can need so the private implementation works in a way that the gfx api is preprocessed and will create a private object for the implementation of opengl/d3d/ogles/etc. this private object will be called by the generic interface. this way there is no virtual call/inheritance and the function inlining is preserved the down side is that each build can only be tied to a single gfx api but that should be ok since you normally ship for a specific platform
there is also a start into the integration of opengles with the power vr sdk. this is very much a work in progress
the build script now has an option to choose a gfx api and some preproc config have been moved in the general generated kinevox config

also a fix for the glew library which is now forced on glfw and glext dependency but there's a better separation between opengl and windriver (even if we depend on glfw, we don't necessarily depend on opengl)

also the opengl format enums have been moved into opengl.h since it is required by both the pbo and the texture

the sprite and the text2d have been converted to vertexbuffer. it builds and run but i haven't verified if it renders

i've also changed the name of the renderer in the tests and made sure that the windows handle is passed. this is not necessary for glfw but is for d3d and gles with egl context

added a new command handle in the winapi windriver )for power saving and screen saver event)

there is also a new pipeline for DLL in case that we depend on libs with dll. it will copy the dll to the executable location so that we can find them at runtime