franky conquers object oriented lua. There used to be a lua module called "object" which was used to access data from the calling object. This was achieved through a "hack" that would store the object pointer in the global before calling the script functions. I made a change where those functions are automatically bounded into the script table so that they are accessible built in. the global pointer to the object has also been moved to the table. this allows us to have a funciton GetActor that fetches the table for that specific actor and then returns it which allows us to call function in a way of self:GetPosition
actors which do not have a scripted component have a table that is auto created when calling GetActor on them which provides the core object functions plus everything in the global.

There is also a new kinevox function called LoadLib which allows us to load a lua library. The inspect lib is an example of this.

we have a new GetName function